fix: allow missing columns in units
When pulling this module into another workflow, it might not always be
the case that you need all of the columns in units.tsv. This PR allows
for missing values, and missing columns in units.tsv. This will prevent
some errors that otherwise would appear regardless of whether or not a
rule needing a certain column is actually used in the end. The
validation of units.tsv is now the responsibility of the workflow
pulling in the module.